+++ recent_visits.user.js
@@ -1,49 +1,33 @@
 // ==UserScript==
 // @name     Recent Visits
-// @version  1.0.1
+// @version  2.0.0
 // @description Various fixes to make the Recent Visits page usable.
 // @include  https://ebird.org/region/*/activity*
 // @include  https://ebird.org/hotspot/*/activity*
 // @require  https://cdnjs.cloudflare.com/ajax/libs/jquery/3.3.1/jquery.js
-// @namespace https://github.com/ProjectBabbler/ebird/
 // @author smackay
-// @copyright 2018 Stuart MacKay (https://github.com/ProjectBabbler/ebird-superscripts)
-// @license MIT
-// @homepage https://github.com/ProjectBabbler/ebird-superscripts
 // @grant GM_addStyle
 // ==/UserScript==
 
-// This script contains all the fixes and changes needed to make the Recent Visits
-// page for a given region or hotspot readable once more. This script contains the
-// functions found in:
-//
-//   recent_visits_blacklist
-//   recent_visits_coordinates
-//   recent_visits_new_tab
-//   recent_visits_restore_visited
-
-
 // Show visited link so you can see what checklists you have read.
-// This performs the same function as the recent_visits_restore_visited script.
 GM_addStyle('td.obstable-date > a:visited { color: darkorange; }');
 
 (function() {
   'use strict';
 
   // Hide all checklists from selected observers.
-  // This performs the same function as the recent_visits_blacklist script.
 
   function hideObserver(name) {
-    $('td[headers="observer"]').each(function () {
+    $('div.Observation-species h3').each(function () {
       if (name ===  $(this).attr('data-observer')) {
-        $(this).closest('tr').addClass('hidden');
+        $(this).closest('li.Observation--placeRecentVisits').css('display', 'none');
       }
     });
   }
 
-  $('td[headers="observer"]').each(function () {
+  $('div.Observation-species h3').each(function () {
     let name = $(this).contents().text().trim().replace(/\s{2,}/, ' ');
-    $(this).prepend('<span style="font-weight: normal; cursor: default" title="Hide all checklists from ' + name + '">x</span> ');
+    $(this).prepend('<span style="font-weight: normal; cursor: default; display: inline-block; margin-right: 5px;" title="Hide all checklists from ' + name + '">x </span> ');
     $(this).attr('data-observer', name);
     $(this).children(":first").click(function () {
       hideObserver(name);
@@ -51,22 +35,21 @@ GM_addStyle('td.obstable-date > a:visited { color: darkorange; }');
   });
 
   // Hide all checklists that contain latitude and longitude in the name.
-  // This performs the same function as the recent_visits_coordinates script.
 
-  const coords = /\-?\d{1,2}[.,]\d{1,4}[,x] ?\-?\d{1,2}[.,]\d{1,4}/;
+  const coords = /\-?\d{1,2}[.,]\d{1,5}[,x] ?\-?\d{1,2}[.,]\d{1,5}/;
 
-  $('td[headers="location"]').each(function () {
+  $('div.Meta--location span.Meta-label').each(function () {
     let name = $(this).contents().text().trim();
     if (coords.test(name)) {
-      $(this).parent().addClass('hidden');
+      $(this).closest('li.Observation--placeRecentVisits').css('display', 'none');
     }
   });
 
   // Open checklists in a new tab view.
-  // This performs the same function as the recent_visits_new_tab script.
 
-  $('td.obstable-date a').each(function () {
+  $('div.Meta--date a.Meta-label').each(function () {
     $(this).attr('target', '_blank');
   });
 
 })();
+